Progress on PE 144.
My initial formula for calculating the slope of the reflected line was
wrong, start from scratch and do the linear algebra without bugs this
time.  Also a lot of reading trying to find a better way but that was
unnecessary - I had the right approach, but with bugs.

Implement slopeOfLine and slopeOfReflectedLine.